Time: 15 ms (19.86%) | Memory: 42.1 MB (97.11%) - LeetSync

+++ 2557-number-of-subarrays-with-lcm-equal-to-k/number-of-subarrays-with-lcm-equal-to-k.java
@@ -21,7 +21,7 @@ class Solution {
         
     }
     public static int lcm(int a,int b){
-        return a*b/gcd(a,b);
+        return (int)( (long)a*b/gcd(a,b));
     }
     public static int gcd(int a,int b){
         return (b==0) ? a:gcd(b,a%b);
